requirements.txt
asgiref>=3.3.1
certifi>=2020.12.5
chardet>=4.0.0
dj-database-url>=0.5.0
Django>=3.1.7
django-colorfield>=0.4.1
django-crispy-forms>=1.11.2
django-environ>=0.4.5
django-js-asset>=1.2.2
django-mptt>=0.12.0
django-mptt-admin>=2.1.0
environ>=1.0
gunicorn>=20.1.0
idna>=2.10
Markdown>=3.3.4
martor>=1.6.0
pytz>=2021.1
requests>=2.25.1
sqlparse>=0.4.1
urllib3>=1.26.4
whitenoise>=5.2.0
mysqlclient
psycopg2-binary
